+++ R/utils.R
@@ -1,19 +1,3 @@
-# Replace logical TRUE and FALSE with character "true" and "false"
-# in a nested list
-replace_true_false <- function(list) {
-  if (is.list(list)) {
-    list <- lapply(list, replace_true_false)
-  } else if (is.logical(list)) {
-    list <- as.character(list)
-    list <- gsub("TRUE", "true", list, fixed = TRUE)
-    list <- gsub("FALSE", "false", list, fixed = TRUE)
-    # Set class of this character vector to 'verbatim'
-    # so that yaml::write_yaml() will not add quotation marks
-    class(list) <- "verbatim"
-  }
-  list
-}
-
 lang_code_chapter_list <- function(chapters_list, language_code) {
   chapters_list <- gsub(
     "\\.Rmd", # nolint: fixed_regex_linter